Updates to ingester to just treat the body as text and regex it.

+++ src/main/scala/models/DocuSignWebhook.scala
@@ -5,7 +5,7 @@ case class DocuSignWebhook(
   apiVersion: String,
   uri: String,
   retryCount: Int,
-  configurationId: Long,
+  configurationId: Int,
   generatedDateTime: String,
   data: WebhookData
 )
